post default thumnail

+++ app/views/dashboard/home.py
@@ -1,6 +1,6 @@
 import os
 from pathlib import Path
-from flask import render_template, request, session, flash, url_for, redirect
+from flask import render_template, request, flash, url_for, redirect, current_app
 from . import dashboard_blueprint
 from app.utils.dashboard_queries.post.post_queries import PostService
 from app.utils.dashboard_queries.category.category_queries import CategoryQueries
@@ -18,13 +18,17 @@ def home():
         if form_data['delete_post']:
             post_id = form_data['delete_post']
             post_to_delete = row_to_dict(post_queries.get_post_by_id(post_id=post_id))
-            # delete post image
+            # if the post has an image
             if post_to_delete['image_url'] is not None or post_to_delete['image_url'] != "":
-                project_root: str = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-                image_path = os.path.join(project_root, post_to_delete['image_url'])
-                # delete post image if exist
-                if Path(image_path).is_file():
-                    os.remove(image_path)
+                # check if it's the default image
+                if post_to_delete['image_url'] != current_app.config["DEFAULT_POST_IMAGE"]:
+                    # if isn't the default image, try to remove it from images/posts folder
+                    project_root: str = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+                    image_path = os.path.join(project_root, post_to_delete['image_url'])
+                    # delete post image if exist
+                    if Path(image_path).is_file():
+                        os.remove(image_path)
+
             deleted, delete_post_msg = post_queries.delete_post(post_id)
             if deleted:
                 flash(delete_post_msg, 'success')
